Moves a test entity
